pyproject.toml
[project]
    classifiers = [
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3",
    ]
    dependencies = [
        "arrow>=1.4.0",
        "babel>=2.17.0",
        "cowsay>=6.1",
        "dotenv>=0.9.9",
        "fastapi>=0.127.0",
        "numpy<2.0.0",
        "onnx-paddleocr>=0.2.1",
        "opencv-python>=4.11.0.86",
        "pandas>=2.3.3",
        "pyyaml>=6.0.3",
        "requests[socks]>=2.32.5",
        "sqlalchemy>=2.0.45",
        "torch>=2.9.1",
        "uvicorn>=0.40.0",
    ]
    description = "My CLI"
    license = { text = "MIT" }
    name = "pymecli"
    readme = "README.md"
    requires-python = ">=3.10"
    version = "0.1.1"

    [project.urls]
        Homepage   = "https://pypi.org/project/pymecli"
        Repository = "https://github.com/meme2046/pymecli"

    [project.scripts]
        bitget = "cli.bitget:app"
        clash  = "cli.clash:cli"
        gate   = "cli.gate:app"
        util   = "cli.util:app"


[tool.uv]
    required-environments = [
        "sys_platform == 'win32' and platform_machine == 'AMD64'",
    ]
    [[tool.uv.index]]
        default = true
        url     = "https://mirrors.aliyun.com/pypi/simple/"

[tool.hatch.build]
    exclude  = []
    packages = ["crypto", "mecli", "utils"]

[build-system]
    build-backend = "hatchling.build"
    requires      = ["hatchling"]

[dependency-groups]
    dev = ["ipykernel", "ruff"]
